requirements.txt
boto3
botocore
awsiotsdk
typer
